+++ Database/Tables/Tracks.sql
@@ -4,7 +4,7 @@
     [Artist_id] VARCHAR(50) NOT NULL,
 	[Album_id] VARCHAR(50) NOT NULL,
 	[disc_number] INT NOT NULL, 
-    [duration_ms] BIGINT NOT NULL, 
+    [duration_ms] INT NOT NULL, 
     [external_urls] VARCHAR(MAX) NOT NULL, 
     [href] VARCHAR(MAX) NOT NULL, 
     [is_local] BIT NOT NULL,
